rag-battle-3/lcpc/requirements.txt
langchain
langchain-community
langchain-openai
openai
pinecone-client
tiktoken
llama-index
torch
transformers
python-pptx
Pillow
docx2txt
unstructured[all-docs]
poppler-utils
pytesseract
python-dotenv